object-file.h: move declarations for object-file.c functions from cache.h
Acked-by: Calvin Wan <calvinwan@google.com>
